requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.12.0
anthropic==0.8.1
redis==5.0.1
loguru==0.7.2
pydantic==2.5.2